fastapi_app/requirements.txt
fastapi
uvicorn
pyjwt
cryptography
